Fix bug in game_over() function
The call to restart() in the game_over() function was wrongly
nested, meaning it would run once for every platform generated
during the game. This made the display choppy after a while.
This commit fixes it. It also makes sure the very first platform
generated is deleted, instead of possibly hanging around in memory
(I have not checked whether it definitely was, but I think it was)