Add support for BGM; mention inefficiency
Godot is nice enough to let a game play if a file is missing,
such as the BGM in this case. So commit the support for it,
even though the music can't be stored in the repo.

Also note the woeful waste of resources, as currently
implemented. Has no actual effect on gameplay, because
we live in the future, but it's messy. So am writing a TODO.